chore: separation of concern
