+++ app/services/price_compare.py
@@ -4,138 +4,28 @@ Price comparison service using Pydantic models for data validation.
 
 import asyncio
 import logging
-from typing import Dict, List, Optional
-from app.utils.constants import GOOGDIT_PRICE_DIVISOR
-from app.models.platform import PlatformType
+
+from app.models.platform import (
+    PlatformType,
+    PlatformConfig,
+    AppediaResponse,
+    MicromazonResponse,
+    GoogditResponse,
+    PriceResult,
+    ComparisonResult,
+)
 
 import httpx
-from pydantic import BaseModel, HttpUrl, Field, field_validator, ConfigDict
+from pydantic import BaseModel, Field, ConfigDict
 
 logger = logging.getLogger(__name__)
-
-
-class PlatformConfig(BaseModel):
-    """Platformm configuration."""
-
-    upc: int = Field(..., gt=0, description="Universal Product Code")
-    url: HttpUrl = Field(..., description="API endpoint URL")
-    platform: PlatformType = Field(..., description="Platform type")
-
-    model_config = ConfigDict(
-        use_enum_values=True,
-    )
-
-
-class AppediaResponse(BaseModel):
-    """API response model."""
-
-    price: str = Field(..., description="Price as string with $ sign")
-    stock: int = Field(..., ge=0, description="Stock quantity")
-
-    @field_validator("price")
-    @classmethod
-    def validate_price(cls, v):
-        """Validate and clean price string."""
-        if not v.startswith("$"):
-            raise ValueError("Price must start with $")
-        return v
-
-    @property
-    def price_float(self) -> float:
-        """Convert price string to float."""
-        return float(self.price.replace("$", "").replace(",", ""))
-
-    @property
-    def in_stock(self) -> bool:
-        """Check if item is in stock."""
-        return self.stock > 0
-
-
-class MicromazonResponse(BaseModel):
-    """Micromazon API response model."""
-
-    available: bool = Field(..., description="Availability status")
-    price: float = Field(..., gt=0, description="Price as float")
-
-    @property
-    def price_float(self) -> float:
-        """Get price as float."""
-        return self.price
-
-    @property
-    def in_stock(self) -> bool:
-        """Check if item is in stock."""
-        return self.available
-
-
-class GoogditLocation(BaseModel):
-    """Googdit location data."""
-
-    l: int = Field(..., description="Location ID")
-    q: int = Field(..., ge=0, description="Quantity available")
-
-
-class GoogditResponse(BaseModel):
-    """Googdit API response model."""
-
-    a: List[GoogditLocation] = Field(..., description="Availability array")
-    p: int = Field(..., gt=0, description="Price in microcents")
-
-    @property
-    def price_float(self) -> float:
-        """Convert microcents to dollars."""
-        return self.p / GOOGDIT_PRICE_DIVISOR
-
-    @property
-    def in_stock(self) -> bool:
-        """Check if item is in stock at any location."""
-        return any(location.q > 0 for location in self.a)
-
-
-class PriceResult(BaseModel):
-    """Standardized price result."""
-
-    platform: PlatformType
-    url: str
-    price: float = Field(..., ge=0)
-    in_stock: bool
-    raw_data: Dict = Field(default_factory=dict)
-    error_message: Optional[str] = None
-
-    @property
-    def is_valid(self) -> bool:
-        """Check if this is a valid result."""
-        return self.error_message is None and self.in_stock and self.price > 0
-
-    def __lt__(self, other):
-        """Enable comparison for finding minimum price."""
-        if not isinstance(other, PriceResult):
-            return NotImplemented
-        return self.price < other.price
-
-    model_config = ConfigDict(use_enum_values=True)
-
-
-class ComparisonResult(BaseModel):
-    """Result of price comparison across platforms."""
-
-    upc: int
-    best_platform: Optional[PlatformType] = None
-    best_price: Optional[float] = None
-    best_url: Optional[str] = None
-    message: str
-    all_results: List[PriceResult] = Field(default_factory=list)
-
-    model_config = ConfigDict(
-        use_enum_values=True,
-    )
 
 
 class PriceComparisonService(BaseModel):
     """Service for comparing prices across multiple platforms using Pydantic models."""
 
     timeout: float = Field(default=10.0, gt=0, description="Request timeout in seconds")
-    platforms: Dict[PlatformType, PlatformConfig] = Field(default_factory=dict)
+    platforms: dict[PlatformType, PlatformConfig] = Field(default_factory=dict)
 
     model_config = ConfigDict(arbitrary_types_allowed=True)
 
@@ -297,7 +187,7 @@ class PriceComparisonService(BaseModel):
         return asyncio.run(self.compare_prices_async(upc))
 
     def _analyze_results(
-        self, upc: int, price_results: List[PriceResult]
+        self, upc: int, price_results: list[PriceResult]
     ) -> ComparisonResult:
         """Analyze price results and find the best option."""
         valid_results = [result for result in price_results if result.is_valid]
